+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Unreleased
+## 4.0.3
 
 - ref: Set wizard version at build time ([#827](https://github.com/getsentry/sentry-wizard/pull/827))
 - fix: Skip asking if users have sentry account if `--org` and `--project` flags were passed ([#817](https://github.com/getsentry/sentry-wizard/pull/817))